+++ src/params/progress.ts
@@ -1,8 +1,0 @@
-export default{
-percentage:     '|number|0|可选，进度条的值最大为100|',
-percentageText:     '|string|--|可选，进度条当前值的文字说明比如：\'30%\'\|\'4/5\'|',
-barbgcolor:     '|string|\'#5170ff\'|可选，进度条的颜色显示，默认为天蓝色|',
-height:     '|string|\'20px\'|可选，进度条的高度值，默认值为20px|',
-isCircle:     '|boolean|false|可选，显示进度条是否为圈形|',
-strokeWidth:     '|number|6|可选，设置圈形进度条宽度，单位是进度条与画布宽度的百分比|',
-};