+++ src/params/status.ts
@@ -1,3 +1,0 @@
-export default{
-type:     '|string|\'invalid\'|必选，类型，值有invalid、running、waiting、important、less-important、error|',
-};