+++ src/params/upload.ts
@@ -1,4 +1,0 @@
-export default{
-preloadFilesRef:     '|TemplateRef<any>|--|可选，用于创建自定义已选择文件列表模板|',
-uploadedFilesRef:     '|TemplateRef<any>|--|可选，用于创建自定义已上传文件列表模板|',
-};